backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
bcrypt>=4.0.0
pyjwt>=2.8.0
